+++ crates/harness-core/src/prompts/mod.rs
@@ -21,10 +21,10 @@ pub use issue::{
     wrap_external_data,
 };
 pub use parsing::{
-    extract_pr_number, extract_review_issues, is_lgtm, is_quota_exhausted, is_waiting,
-    parse_complexity, parse_created_issue_number, parse_github_pr_url, parse_issue_count,
-    parse_plan_issue, parse_pr_url, parse_triage, repo_slug_from_pr_url, TriageComplexity,
-    TriageDecision,
+    extract_pr_number, extract_review_issues, has_gemini_body_feedback, is_lgtm,
+    is_quota_exhausted, is_waiting, parse_complexity, parse_created_issue_number,
+    parse_github_pr_url, parse_issue_count, parse_plan_issue, parse_pr_url, parse_triage,
+    repo_slug_from_pr_url, TriageComplexity, TriageDecision,
 };
 pub use pr::{check_existing_pr, continue_existing_pr, rebase_conflicting_pr};
 pub use review::{
